Debug: Use global integer precision constant in pricer.
Previously it was a wrong pricer stopping criteria which is that reduced cost < -1, now it has been changed to reduced cost < -integer precision. Criteria for whether to add a column or not is also updated with this constant. Previously -1e8 is used and it caused unknown bug which leads to D-W reformulation stopped at a value which is higher than the actual optimal